Tier2: fix regression preventing remote errors from being logged
When CURLParams::failOnError is set, it makes code stop instantly when the request has errored somewhere (400 or higher response codes), but this prevents code from displaying the error reason attached to the failed request.

Setting this to false by default fixes the problem; if this option needs to be used it should be set explicitly.