ImguiSystem: fix input latency caused by mutex
There were reports of high input latency even though we implemented NVIDIA Reflex and AMD Anti-Lag 2.
The reason for this increased latency was because CThreadFastMutex was being used in the window procedure instead of a CThread mutex.

The contention for CImguiSystem::m_systemInitState in CImguiSystem::MessageHandler() is high, and therefore we should not use CThreadFastMutex as its only effective  in its fastness when the contention isn't high. Using CThreadMutex instead resulted in an extreme reduction of input latency.