MaterialSystem: don't sample imgui frames if window is occluded
Check if the window is occluded first before sampling. Massive runtime load reduction when application minimized in full-screen mode.

This also fixes the issue where the console would scroll all the way to the top when the application is minimized in full screen mode, as we can't calculate the window sizes properly causing SetScrollPos to set the pos to a large negative value.