MaterialSystem: only update STBSP camera if we aren't GPU driven
Only update the stream camera if we aren't GPU driven to save on load in the render thread.

The camera is used to get the correct column from the STBSP file. Though, if we use the GPU feedback driven texture streaming system, it doesn't make sense to burn CPU here since the GPU system doesn't use the camera at all.