MaterialSystem: only run model texture crediting if we aren't GPU driven
The compute shader already takes care of this. This code updates the STBSP histogram which won't be used if the cvar 'gpu_driven_tex_stream' is set.

Added a check for if 'gpu_driven_tex_stream' is set, and if so, we don't run this code to save on CPU load.