MaterialSystem: fix incorrect vendor constant for AMD graphics hardware
Users have reported that AMD Anti-Lag 2 didn't work. The issue turned out to be the incorrect vendor constant. After this patch the Anti-Lag 2 implementation appears to work perfectly.

This constant is used to check if installed (and currently selected) hardware is of target vendor before making any driver API calls.