Tier1: properly handle null string values in ConVar::InternalSetValue
There is code that checks if the given string is nullptr, and if so, sets it to an empty string (""). But this new pointer wasn't used on InternalSetColorFromString and atof. InternalSetColorFromString calls sscanf on the given string which if null, is undefined behavior.

The bug was also present in the engine code, this has been patched on assembly level and confirmed correct.